requirements.txt
fastapi 
pydantic 
mysql-connector-python 
pyjwt 
uvicorn
